+++ TLM/TLM/State/TMPEMoveITIntegration.cs
@@ -4,6 +4,7 @@ namespace TrafficManager.State {
     using System.Collections.Generic;
     using TrafficManager.Util.Record;
     using CSUtil.Commons;
+    using TrafficManager.Lifecycle;
 
     public class TMPEMoveItIntegrationFactory : IMoveItIntegrationFactory {
         public MoveItIntegrationBase GetInstance() => new TMPEMoveItIntegration();
