+++ TLM/TLM/UI/MainMenu/MainMenuWindow.cs
@@ -14,6 +14,7 @@ namespace TrafficManager.UI.MainMenu {
     using TrafficManager.U.Autosize;
     using TrafficManager.Util;
     using UnityEngine;
+    using TrafficManager.Lifecycle;
 
     public class MainMenuWindow
         : U.Panel.BaseUWindowPanel,
@@ -372,7 +373,7 @@ namespace TrafficManager.UI.MainMenu {
                 }
             }
 
-            if (TrafficManagerMod.InGameHotReload) {
+            if (TMPELifecycle.Instance.InGameHotReload) {
                 // Hot Reload version label (debug only)
                 string text = $"HOT RELOAD {Assembly.GetExecutingAssembly().GetName().Version}";
                 using (var hotReloadB = builder.Label<U.ULabel>(text)) {
