+++ TLM/TLM/UI/TrafficManagerTool.cs
@@ -27,6 +27,7 @@ namespace TrafficManager.UI {
     using TrafficManager.UI.SubTools.LaneArrows;
     using TrafficManager.UI.SubTools.PrioritySigns;
     using TrafficManager.UI.SubTools.TimedTrafficLights;
+    using TrafficManager.Lifecycle;
 
     using static TrafficManager.Util.Shortcuts;
     using static TrafficManager.Util.SegmentTraverser;
@@ -244,7 +245,7 @@ namespace TrafficManager.UI {
         public void SetToolMode(ToolMode newToolMode) {
             ToolMode oldToolMode = toolMode_;
 
-            if(toolMode_ != ToolMode.None && LoadingExtension.PlayMode) {
+            if(toolMode_ != ToolMode.None && TMPELifecycle.PlayMode) {
                 // Make it impossible for user to undo changes performed by Road selection panels
                 // after changing traffic rule vis other tools.
                 // TODO: This code will not be necessary when we implement intent.
@@ -377,7 +378,7 @@ namespace TrafficManager.UI {
         /// </summary>
         void DefaultRenderOverlay(RenderManager.CameraInfo cameraInfo)
         {
-            if (!LoadingExtension.PlayMode) {
+            if (!TMPELifecycle.PlayMode) {
                 return; // world info view panels are not availble in edit mode
             }
             SubTools.PrioritySigns.MassEditOverlay.Show
@@ -552,7 +553,7 @@ namespace TrafficManager.UI {
         }
 
         void DefaultOnToolGUI(Event e) {
-            if (!LoadingExtension.PlayMode) {
+            if (!TMPELifecycle.PlayMode) {
                 return; // world info view panels are not availble in edit mode
             }
             if (e.type == EventType.MouseDown && e.button == 0) {
@@ -990,7 +991,7 @@ namespace TrafficManager.UI {
         /// <summary>Shows a tutorial message. Must be called by a Unity thread.</summary>
         /// <param name="localeKey">Tutorial key.</param>
         public static void ShowAdvisor(string localeKey) {
-            if (!GlobalConfig.Instance.Main.EnableTutorial || !LoadingExtension.PlayMode) {
+            if (!GlobalConfig.Instance.Main.EnableTutorial || !TMPELifecycle.PlayMode) {
                 return;
             }
